Simplified pipelining
